+++ test/test.dart
@@ -114,6 +114,23 @@ class _AllTypes {
   late ObjectId objectIdProp;
   late Uuid uuidProp;
   late int intProp;
+
+  late String? nullableStringProp;
+  late bool? nullableBoolProp;
+  late DateTime? nullableDateProp;
+  late double? nullableDoubleProp;
+  late ObjectId? nullableObjectIdProp;
+  late Uuid? nullableUuidProp;
+  late int? nullableIntProp;
+}
+
+@RealmModel()
+class _LinksClass {
+  @PrimaryKey()
+  late Uuid id;
+
+  late _LinksClass? link;
+  late List<_LinksClass> list;
 }
 
 @RealmModel()